+++ Changes.md
@@ -2,5 +2,6 @@
 
 ## NEXT
 
+* Add Nightwatch.js-based tests.
 * README markdown fixes. [GH#57, olleolleolle]